requirements.txt
Django==5.1.6
celery==5.4.0
pillow==11.2.1
psycopg2_binary==2.9.10
pydantic-settings==2.7.1
django-celery-results==2.5.1
django-celery-beat==2.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
django-filter==25.1
drf-spectacular==0.28.0
django-cors-headers==4.7.0
requests==2.32.3
redis==5.2.1
boto3==1.35.95
django-storages==1.14.6